+++ NEWS.md
@@ -12,6 +12,10 @@ title: "NEWS for the emmeans package"
   * Added alt text to all pictures in vignettes (#389). This makes
     the materials more accessible per guidelines from the 
     [A11Y project](https://www.a11yproject.com/).
+  * Added `"atanh"` to the options in `make.tran()`
+  * `make.tran()` replaces `param` argument with `alpha` and `beta`
+    (`param` is still supported for backward compatibility)
+    and documentation has been revised in hopes of making everything clearer
     
 
 ## emmeans 1.8.2